refactor redux
