+++ frontend/src/actionCreators/gameState/creators.ts
@@ -1,23 +1,23 @@
 import { GameStateActions } from '@frontend/src/actionCreators/gameState/actions';
-import { GameState } from '@frontend/src/reducers/gameState';
-import { PayloadAction } from '@frontend/src/actionCreators/models/action';
+import { defaultGameState, GameStateAction } from '@frontend/src/reducers/gameState';
 
-export function createUpdateMyLifePercent(gameState: GameState): PayloadAction<GameStateActions, GameState> {
-  return {
+export const setMyLifePercent = (myLifePercent: number): Partial<GameStateAction> => (
+  {
     type: GameStateActions.UPDATE_MY_LIFE_PERCENT,
-    payload: gameState,
-  };
-}
-export function createUpdateEnemyLifePercent(gameState: GameState): PayloadAction<GameStateActions, GameState> {
-  return {
-    type: GameStateActions.UPDATE_ENEMY_LIFE_PERCENT,
-    payload: gameState,
-  };
-}
+    payload: { ...defaultGameState, myLifePercent },
+  }
+);
 
-export function createUpdateRoundName(gameState: GameState): PayloadAction<GameStateActions, GameState> {
-  return {
+export const setEnemyLifePercent = (enemyLifePercent: number): Partial<GameStateAction> => (
+  {
+    type: GameStateActions.UPDATE_ENEMY_LIFE_PERCENT,
+    payload: { ...defaultGameState, enemyLifePercent },
+  }
+);
+
+export const setRoundName = (roundName: string): Partial<GameStateAction> => (
+  {
     type: GameStateActions.UPDATE_ROUND_NAME,
-    payload: gameState,
-  };
-}
+    payload: { ...defaultGameState, roundName },
+  }
+);
